[Proposal] Add async capability for celery
